+++ c/tx/source/cff.h
@@ -10,7 +10,7 @@
 "+/-d    enable duplicate hintsubs warnings\n"
 "-n      remove hints\n"
 "-std    force the output font to have StandardEncoding\n"
-"-no_futile remove futile subrs during subroutinization\n"
+"-no_futile remove futile subrs during subroutinization (deprecated; always in effect)\n"
 "\n"
 "CFF mode writes a CFF conversion of an abstract font. The precise form of the\n"
 "CFF font that is written can be controlled to a limited extent by the options\n",